Change Ids to Long; update products view

+++ src/main/java/com/example/demo/repositories/ApplicationUserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.example.demo.entities.ApplicationUser;
 
-public interface ApplicationUserRepository extends CrudRepository<ApplicationUser, Integer> {
+public interface ApplicationUserRepository extends CrudRepository<ApplicationUser, Long> {
 	
 	Optional<ApplicationUser> findByUsername(String username);
 }
